Updated SEO for entire blog site.
Added two different SEO types, one for the blog posts and one for
the rest of the pages. The first accepts all of the relevant props
for the blog and the second takes the prop of title and description.
The package that I used for all of the SEO optimizations is called
Next-Seo.