Changed author layout on posts & added Ayesha
Rather than displaying a link to the author's twitter, I now display a
link to the author's web page because it is more relevant.

This is done by modifying the post layout component to use the author's
website entry that will be located within the author's information
in their respective profiles.

I also added Ayesha as an Author.